Upgrading packages for Python 3 + npm -> yarn
Python 2 is EOL after this year, so updating to the latest packages to
support at least Python 3.7. App still doesn't work fully yet, but at
least it runs.

Also included a change from npm to yarn since yarn is included in the
latest version of NodeJS (and yarn is the better tool nowadays).