Added pyenv instructions + a bash script to start the API and front end
I'd rather see yarn run the API and front end in parallel, but that can
be for later. This is keeping the status quo, but just making it simpler
to restart when needed.

I'm including pyenv in the readme as it makes setting up the python
versions easier if you don't have it already.